# backend/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/graphic/all/": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get All With Pagination",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Graphic"
                ],
                "summary": "Get All With Pagination",
                "parameters": [
                    {
                        "description": "Get All With Pagination",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PaginationInputWithFilter"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Graphic response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/v1/graphic/brand/{brand}": {
            "get": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get All With Brand filter",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Graphic"
                ],
                "summary": "Get All With Brand filter",
                "responses": {
                    "201": {
                        "description": "Graphic response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/v1/graphic/create": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Create a Graphic",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Graphic"
                ],
                "summary": "Create a Graphic",
                "parameters": [
                    {
                        "description": "Create a Graphic",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateGraphicRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Graphic response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/v1/graphic/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Delete Graphic",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Graphic"
                ],
                "summary": "Delete Graphic",
                "responses": {
                    "204": {
                        "description": "Graphic response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/v1/graphic/get/{id}": {
            "get": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "GetGraphicById",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Graphic"
                ],
                "summary": "Get Graphic By Id",
                "responses": {
                    "200": {
                        "description": "GetGraphicById response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/v1/graphic/update/{id}": {
            "put": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Update Graphic",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Graphic"
                ],
                "summary": "Update Graphic",
                "parameters": [
                    {
                        "description": " Update Graphic",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateGraphicRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Graphic response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/v1/hdd/all/": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get All With Pagination",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "HardDrive"
                ],
                "summary": "Get All With Pagination",
                "parameters": [
                    {
                        "description": "Get All With Pagination",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PaginationInputWithFilter"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "HardDrive response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/v1/hdd/create": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Create a HardDrive",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "HardDrive"
                ],
                "summary": "Create a HardDrive",
                "parameters": [
                    {
                        "description": "Create a HardDrive",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateHardDriveRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "HardDrive response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/v1/hdd/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Delete HardDrive",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "HardDrive"
                ],
                "summary": "Delete HardDrive",
                "responses": {
                    "204": {
                        "description": "HardDrive response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/v1/hdd/get/{id}": {
            "get": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "GetHardDriveById",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "HardDrive"
                ],
                "summary": "Get HardDrive By Id",
                "responses": {
                    "200": {
                        "description": "GetHardDriveById response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/v1/hdd/update/{id}": {
            "put": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Update HardDrive",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "HardDrive"
                ],
                "summary": "Update HardDrive",
                "parameters": [
                    {
                        "description": " Update HardDrive",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateHardDriveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HardDrive response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.CreateGraphicRequest": {
            "type": "object",
            "properties": {
                "gpuBrandId": {
                    "type": "integer"
                },
                "gpuName": {
                    "type": "string"
                },
                "powerConsumption": {
                    "type": "number"
                }
            }
        },
        "dto.CreateHardDriveRequest": {
            "type": "object",
            "properties": {
                "powerConsumption": {
                    "type": "number"
                },
                "rpm": {
                    "type": "integer"
                },
                "size": {
                    "type": "number"
                }
            }
        },
...
        "dto.Filter": {
            "type": "object",
            "properties": {
                "filterType": {
                    "type": "string"
                },
                "form": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "dto.PaginationInputWithFilter": {
            "type": "object",
            "properties": {
                "filter": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/dto.Filter"
                    }
                },
                "page_number": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "sort": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SortDto"
                    }
                }
            }
        },
...
        "dto.SortDto": {
            "type": "object",
            "properties": {
                "colId": {
                    "type": "string"
                },
                "string": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateGraphicRequest": {
            "type": "object",
            "properties": {
                "gpuName": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateHardDriveRequest": {
            "type": "object",
            "properties": {
                "rpm": {
                    "type": "integer"
                },
                "size": {
                    "type": "number"
                }
            }
        },
...
        "helper.Response": {
            "type": "object",
            "properties": {
                "err": {
                    "type": "string"
                },
                "result": {},
                "statusCode": {
                    "$ref": "#/definitions/helper.ResultCode"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "helper.ResultCode": {
            "type": "integer",
            "enum": [
                0,
                40001,
                40002,
                40101,
                40301,
                40401,
                42901,
                42902,
                50001,
                50002
            ],
            "x-enum-varnames": [
                "Success",
                "ValidationError",
                "Error",
                "AuthError",
                "ForbiddenError",
                "NotFoundError",
                "LimiterError",
                "OtpLimiterError",
                "CustomRecovery",
                "InternalError"
            ]
        }
    }
}